unify embedding tests
