+++ flair/__init__.py
@@ -9,7 +9,7 @@ from transformers import set_seed as hf_set_seed
 cache_root = Path(os.getenv("FLAIR_CACHE_ROOT", Path(Path.home(), ".flair")))
 
 # global variable: device
-if torch.cuda.is_available() and False:
+if torch.cuda.is_available():
     device = torch.device("cuda:0")
 else:
     device = torch.device("cpu")
